Address pages: add epoch data
