+++ apps/block_scout_web/lib/block_scout_web/controllers/address_celo_controller.ex
@@ -6,6 +6,7 @@ defmodule BlockScoutWeb.AddressCeloController do
   import BlockScoutWeb.Account.AuthController, only: [current_user: 1]
   import BlockScoutWeb.Models.GetAddressTags, only: [get_address_tags: 2]
 
+  alias Explorer.Celo.EpochUtil
   alias Explorer.{Chain, Market}
   alias Explorer.ExchangeRates.Token
   alias Indexer.Fetcher.CoinBalanceOnDemand
@@ -16,6 +17,9 @@ defmodule BlockScoutWeb.AddressCeloController do
     with {:ok, address_hash} <- Chain.string_to_address_hash(address_hash_string),
          {:ok, address} <- Chain.hash_to_address(address_hash),
          %CeloAccount{address: _} <- address.celo_account do
+      {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+        EpochUtil.get_address_summary(address)
+
       Logger.debug("Parsing Celo Address #{address_hash_string}")
 
       render(
@@ -26,7 +30,12 @@ defmodule BlockScoutWeb.AddressCeloController do
         coin_balance_status: CoinBalanceOnDemand.trigger_fetch(address),
         exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
         counters_path: address_path(conn, :address_counters, %{"id" => address_hash_string}),
-        tags: get_address_tags(address_hash, current_user(conn))
+        tags: get_address_tags(address_hash, current_user(conn)),
+        validator_or_group_sum: validator_or_group_sum,
+        voting_sum: voting_sum,
+        locked_gold: locked_gold,
+        vote_activated_gold: vote_activated_gold,
+        pending_gold: pending_gold
       )
     else
       _ ->
