+++ apps/block_scout_web/lib/block_scout_web/controllers/address_coin_balance_controller.ex
@@ -10,6 +10,7 @@ defmodule BlockScoutWeb.AddressCoinBalanceController do
   import BlockScoutWeb.Models.GetAddressTags, only: [get_address_tags: 2]
 
   alias BlockScoutWeb.{AccessHelpers, AddressCoinBalanceView, Controller}
+  alias Explorer.Celo.EpochUtil
   alias Explorer.{Chain, Market}
   alias Explorer.Chain.{Address, Wei}
   alias Explorer.ExchangeRates.Token
@@ -73,13 +74,21 @@ defmodule BlockScoutWeb.AddressCoinBalanceController do
     with {:ok, address_hash} <- Chain.string_to_address_hash(address_hash_string),
          {:ok, address} <- Chain.hash_to_address(address_hash),
          {:ok, false} <- AccessHelpers.restricted_access?(address_hash_string, params) do
+      {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+        EpochUtil.get_address_summary(address)
+
       render(conn, "index.html",
         address: address,
         coin_balance_status: CoinBalanceOnDemand.trigger_fetch(address),
         exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
         current_path: Controller.current_full_path(conn),
         counters_path: address_path(conn, :address_counters, %{"id" => Address.checksum(address_hash)}),
-        tags: get_address_tags(address_hash, current_user(conn))
+        tags: get_address_tags(address_hash, current_user(conn)),
+        validator_or_group_sum: validator_or_group_sum,
+        voting_sum: voting_sum,
+        locked_gold: locked_gold,
+        vote_activated_gold: vote_activated_gold,
+        pending_gold: pending_gold
       )
     else
       {:restricted_access, _} ->
@@ -97,6 +106,9 @@ defmodule BlockScoutWeb.AddressCoinBalanceController do
 
         case Chain.Hash.Address.validate(address_hash_string) do
           {:ok, _} ->
+            {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+              EpochUtil.get_address_summary(address)
+
             render(
               conn,
               "index.html",
@@ -105,7 +117,12 @@ defmodule BlockScoutWeb.AddressCoinBalanceController do
               exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
               counters_path: address_path(conn, :address_counters, %{"id" => Address.checksum(address_hash)}),
               current_path: Controller.current_full_path(conn),
-              tags: get_address_tags(address_hash, current_user(conn))
+              tags: get_address_tags(address_hash, current_user(conn)),
+              validator_or_group_sum: validator_or_group_sum,
+              voting_sum: voting_sum,
+              locked_gold: locked_gold,
+              vote_activated_gold: vote_activated_gold,
+              pending_gold: pending_gold
             )
 
           _ ->
