+++ apps/block_scout_web/lib/block_scout_web/controllers/address_contract_controller.ex
@@ -8,6 +8,7 @@ defmodule BlockScoutWeb.AddressContractController do
 
   alias BlockScoutWeb.AccessHelpers
   alias BlockScoutWeb.AddressContractVerificationViaJsonController, as: VerificationController
+  alias Explorer.Celo.EpochUtil
   alias Explorer.{Chain, Market}
   alias Explorer.Etherscan.Contracts
   alias Explorer.ExchangeRates.Token
@@ -29,6 +30,9 @@ defmodule BlockScoutWeb.AddressContractController do
          {:ok, false} <- AccessHelpers.restricted_access?(address_hash_string, params),
          _ <- VerificationController.check_and_verify(address_hash_string),
          {:ok, address} <- Chain.find_contract_address(address_hash, address_options, true) do
+      {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+        EpochUtil.get_address_summary(address)
+
       with {:ok, implementation_address} <- Contracts.get_proxied_address(address_hash),
            {:ok, implementation_contract} <- Chain.find_contract_address(implementation_address, address_options, true) do
         Logger.debug("Implementation address FOUND in proxy table #{implementation_address}")
@@ -43,7 +47,12 @@ defmodule BlockScoutWeb.AddressContractController do
           coin_balance_status: CoinBalanceOnDemand.trigger_fetch(address),
           exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
           counters_path: address_path(conn, :address_counters, %{"id" => address_hash_string}),
-          tags: get_address_tags(address_hash, current_user(conn))
+          tags: get_address_tags(address_hash, current_user(conn)),
+          validator_or_group_sum: validator_or_group_sum,
+          voting_sum: voting_sum,
+          locked_gold: locked_gold,
+          vote_activated_gold: vote_activated_gold,
+          pending_gold: pending_gold
         )
       else
         {:error, :not_found} ->
@@ -58,7 +67,12 @@ defmodule BlockScoutWeb.AddressContractController do
             coin_balance_status: CoinBalanceOnDemand.trigger_fetch(address),
             exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
             counters_path: address_path(conn, :address_counters, %{"id" => address_hash_string}),
-            tags: get_address_tags(address_hash, current_user(conn))
+            tags: get_address_tags(address_hash, current_user(conn)),
+            validator_or_group_sum: validator_or_group_sum,
+            voting_sum: voting_sum,
+            locked_gold: locked_gold,
+            vote_activated_gold: vote_activated_gold,
+            pending_gold: pending_gold
           )
       end
     else
