+++ apps/block_scout_web/lib/block_scout_web/controllers/address_decompiled_contract_controller.ex
@@ -5,6 +5,7 @@ defmodule BlockScoutWeb.AddressDecompiledContractController do
   import BlockScoutWeb.Models.GetAddressTags, only: [get_address_tags: 2]
 
   alias BlockScoutWeb.AccessHelpers
+  alias Explorer.Celo.EpochUtil
   alias Explorer.{Chain, Market}
   alias Explorer.ExchangeRates.Token
   alias Indexer.Fetcher.CoinBalanceOnDemand
@@ -13,6 +14,9 @@ defmodule BlockScoutWeb.AddressDecompiledContractController do
     with {:ok, address_hash} <- Chain.string_to_address_hash(address_hash_string),
          {:ok, address} <- Chain.find_decompiled_contract_address(address_hash),
          {:ok, false} <- AccessHelpers.restricted_access?(address_hash_string, params) do
+      {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+        EpochUtil.get_address_summary(address)
+
       render(
         conn,
         "index.html",
@@ -20,7 +24,12 @@ defmodule BlockScoutWeb.AddressDecompiledContractController do
         coin_balance_status: CoinBalanceOnDemand.trigger_fetch(address),
         exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
         counters_path: address_path(conn, :address_counters, %{"id" => address_hash_string}),
-        tags: get_address_tags(address_hash, current_user(conn))
+        tags: get_address_tags(address_hash, current_user(conn)),
+        validator_or_group_sum: validator_or_group_sum,
+        voting_sum: voting_sum,
+        locked_gold: locked_gold,
+        vote_activated_gold: vote_activated_gold,
+        pending_gold: pending_gold
       )
     else
       {:restricted_access, _} ->
