+++ apps/block_scout_web/lib/block_scout_web/controllers/address_epoch_transaction_controller.ex
@@ -10,8 +10,9 @@ defmodule BlockScoutWeb.AddressEpochTransactionController do
   import BlockScoutWeb.Models.GetAddressTags, only: [get_address_tags: 2]
 
   alias BlockScoutWeb.{AccessHelpers, Controller, EpochTransactionView}
+  alias Explorer.Celo.EpochUtil
   alias Explorer.{Chain, Market}
-  alias Explorer.Chain.{CeloAccountEpoch, CeloElectionRewards, Wei}
+  alias Explorer.Chain.CeloElectionRewards
   alias Explorer.ExchangeRates.Token
   alias Indexer.Fetcher.CoinBalanceOnDemand
   alias Phoenix.View
@@ -60,12 +61,8 @@ defmodule BlockScoutWeb.AddressEpochTransactionController do
     with {:ok, address_hash} <- Chain.string_to_address_hash(address_hash_string),
          {:ok, address} <- Chain.hash_to_address(address_hash),
          {:ok, false} <- AccessHelpers.restricted_access?(address_hash_string, params) do
-      {validator_or_group_sum, voting_sum} = get_sums(address)
-
-      last_account_epoch = CeloAccountEpoch.last_for_address(address_hash)
-      {locked_gold, vote_activated_gold} = last_account_epoch |> calculate_locked_and_vote_activated_gold()
-
-      pending_gold = Chain.fetch_sum_available_celo_unlocked_for_address(address_hash)
+      {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+        EpochUtil.get_address_summary(address)
 
       render(
         conn,
@@ -95,14 +92,6 @@ defmodule BlockScoutWeb.AddressEpochTransactionController do
     end
   end
 
-  defp calculate_locked_and_vote_activated_gold(nil) do
-    {:ok, zero_wei} = Wei.cast(0)
-    {zero_wei, zero_wei}
-  end
-
-  defp calculate_locked_and_vote_activated_gold(account_epoch),
-    do: {account_epoch.total_locked_gold, Wei.sub(account_epoch.total_locked_gold, account_epoch.nonvoting_locked_gold)}
-
   defp get_rewards(%Chain.Address{celo_account: nil, hash: address_hash}, params) do
     CeloElectionRewards.get_paginated_rewards_for_address([address_hash], ["delegated_payment"], params)
   end
@@ -118,17 +107,6 @@ defmodule BlockScoutWeb.AddressEpochTransactionController do
           [type, "voter", "delegated_payment"],
           params
         )
-    end
-  end
-
-  defp get_sums(%Chain.Address{celo_account: nil, hash: address_hash}) do
-    {nil, CeloElectionRewards.get_rewards_sum_for_account(address_hash)}
-  end
-
-  defp get_sums(address) do
-    case address.celo_account.account_type do
-      "normal" -> {nil, CeloElectionRewards.get_rewards_sum_for_account(address.hash)}
-      type -> CeloElectionRewards.get_rewards_sums_for_account(address.hash, type)
     end
   end
 end
