+++ apps/block_scout_web/lib/block_scout_web/controllers/address_internal_transaction_controller.ex
@@ -10,6 +10,7 @@ defmodule BlockScoutWeb.AddressInternalTransactionController do
   import BlockScoutWeb.Models.GetAddressTags, only: [get_address_tags: 2]
 
   alias BlockScoutWeb.{AccessHelpers, Controller, InternalTransactionView}
+  alias Explorer.Celo.EpochUtil
   alias Explorer.{Chain, Market}
   alias Explorer.Chain.{Address, Wei}
   alias Explorer.ExchangeRates.Token
@@ -80,6 +81,9 @@ defmodule BlockScoutWeb.AddressInternalTransactionController do
     with {:ok, address_hash} <- Chain.string_to_address_hash(address_hash_string),
          {:ok, address} <- Chain.hash_to_address(address_hash),
          {:ok, false} <- AccessHelpers.restricted_access?(address_hash_string, params) do
+      {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+        EpochUtil.get_address_summary(address)
+
       render(
         conn,
         "index.html",
@@ -89,7 +93,12 @@ defmodule BlockScoutWeb.AddressInternalTransactionController do
         exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
         filter: params["filter"],
         counters_path: address_path(conn, :address_counters, %{"id" => address_hash_string}),
-        tags: get_address_tags(address_hash, current_user(conn))
+        tags: get_address_tags(address_hash, current_user(conn)),
+        validator_or_group_sum: validator_or_group_sum,
+        voting_sum: voting_sum,
+        locked_gold: locked_gold,
+        vote_activated_gold: vote_activated_gold,
+        pending_gold: pending_gold
       )
     else
       {:restricted_access, _} ->
@@ -107,6 +116,9 @@ defmodule BlockScoutWeb.AddressInternalTransactionController do
 
         case Chain.Hash.Address.validate(address_hash_string) do
           {:ok, _} ->
+            {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+              EpochUtil.get_address_summary(address)
+
             render(
               conn,
               "index.html",
@@ -116,7 +128,12 @@ defmodule BlockScoutWeb.AddressInternalTransactionController do
               exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
               counters_path: address_path(conn, :address_counters, %{"id" => Address.checksum(address_hash)}),
               current_path: Controller.current_full_path(conn),
-              tags: get_address_tags(address_hash, current_user(conn))
+              tags: get_address_tags(address_hash, current_user(conn)),
+              validator_or_group_sum: validator_or_group_sum,
+              voting_sum: voting_sum,
+              locked_gold: locked_gold,
+              vote_activated_gold: vote_activated_gold,
+              pending_gold: pending_gold
             )
 
           _ ->
