+++ apps/block_scout_web/lib/block_scout_web/controllers/address_transaction_controller.ex
@@ -12,6 +12,7 @@ defmodule BlockScoutWeb.AddressTransactionController do
   import BlockScoutWeb.Models.GetAddressTags, only: [get_address_tags: 2]
 
   alias BlockScoutWeb.{AccessHelpers, Controller, TransactionView}
+  alias Explorer.Celo.EpochUtil
   alias Explorer.{Chain, Market}
   alias Explorer.Export.CSV
 
@@ -117,6 +118,9 @@ defmodule BlockScoutWeb.AddressTransactionController do
     with {:ok, address_hash} <- Chain.string_to_address_hash(address_hash_string),
          {:ok, address} <- Chain.hash_to_address(address_hash),
          {:ok, false} <- AccessHelpers.restricted_access?(address_hash_string, params) do
+      {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+        EpochUtil.get_address_summary(address)
+
       render(
         conn,
         "index.html",
@@ -126,7 +130,12 @@ defmodule BlockScoutWeb.AddressTransactionController do
         filter: params["filter"],
         counters_path: address_path(conn, :address_counters, %{"id" => address_hash_string}),
         current_path: Controller.current_full_path(conn),
-        tags: get_address_tags(address_hash, current_user(conn))
+        tags: get_address_tags(address_hash, current_user(conn)),
+        validator_or_group_sum: validator_or_group_sum,
+        voting_sum: voting_sum,
+        locked_gold: locked_gold,
+        vote_activated_gold: vote_activated_gold,
+        pending_gold: pending_gold
       )
     else
       :error ->
@@ -147,6 +156,9 @@ defmodule BlockScoutWeb.AddressTransactionController do
 
         case Chain.Hash.Address.validate(address_hash_string) do
           {:ok, _} ->
+            {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+              EpochUtil.get_address_summary(address)
+
             render(
               conn,
               "index.html",
@@ -156,7 +168,12 @@ defmodule BlockScoutWeb.AddressTransactionController do
               filter: params["filter"],
               counters_path: address_path(conn, :address_counters, %{"id" => address_hash_string}),
               current_path: Controller.current_full_path(conn),
-              tags: get_address_tags(address_hash, current_user(conn))
+              tags: get_address_tags(address_hash, current_user(conn)),
+              validator_or_group_sum: validator_or_group_sum,
+              voting_sum: voting_sum,
+              locked_gold: locked_gold,
+              vote_activated_gold: vote_activated_gold,
+              pending_gold: pending_gold
             )
 
           _ ->
