+++ apps/block_scout_web/lib/block_scout_web/controllers/address_write_contract_controller.ex
@@ -13,6 +13,7 @@ defmodule BlockScoutWeb.AddressWriteContractController do
 
   alias BlockScoutWeb.AccessHelpers
   alias BlockScoutWeb.AddressView
+  alias Explorer.Celo.EpochUtil
   alias Explorer.{Chain, Market}
   alias Explorer.Chain.Address
   alias Explorer.ExchangeRates.Token
@@ -42,6 +43,9 @@ defmodule BlockScoutWeb.AddressWriteContractController do
          {:ok, address} <- Chain.find_contract_address(address_hash, address_options, true),
          false <- is_nil(address.smart_contract),
          {:ok, false} <- AccessHelpers.restricted_access?(address_hash_string, params) do
+      {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+        EpochUtil.get_address_summary(address)
+
       render(
         conn,
         "index.html",
@@ -51,7 +55,12 @@ defmodule BlockScoutWeb.AddressWriteContractController do
             non_custom_abi: true,
             coin_balance_status: CoinBalanceOnDemand.trigger_fetch(address),
             counters_path: address_path(conn, :address_counters, %{"id" => Address.checksum(address_hash)}),
-            tags: get_address_tags(address_hash, current_user(conn))
+            tags: get_address_tags(address_hash, current_user(conn)),
+            validator_or_group_sum: validator_or_group_sum,
+            voting_sum: voting_sum,
+            locked_gold: locked_gold,
+            vote_activated_gold: vote_activated_gold,
+            pending_gold: pending_gold
           ]
       )
     else
@@ -60,6 +69,9 @@ defmodule BlockScoutWeb.AddressWriteContractController do
           with {:ok, address_hash} <- Chain.string_to_address_hash(address_hash_string),
                {:ok, address} <- Chain.find_contract_address(address_hash, address_options, false),
                {:ok, false} <- AccessHelpers.restricted_access?(address_hash_string, params) do
+            {validator_or_group_sum, voting_sum, locked_gold, vote_activated_gold, pending_gold} =
+              EpochUtil.get_address_summary(address)
+
             render(
               conn,
               "index.html",
@@ -69,7 +81,12 @@ defmodule BlockScoutWeb.AddressWriteContractController do
                   non_custom_abi: false,
                   coin_balance_status: CoinBalanceOnDemand.trigger_fetch(address),
                   counters_path: address_path(conn, :address_counters, %{"id" => Address.checksum(address_hash)}),
-                  tags: get_address_tags(address_hash, current_user(conn))
+                  tags: get_address_tags(address_hash, current_user(conn)),
+                  validator_or_group_sum: validator_or_group_sum,
+                  voting_sum: voting_sum,
+                  locked_gold: locked_gold,
+                  vote_activated_gold: vote_activated_gold,
+                  pending_gold: pending_gold
                 ]
             )
           else
