+++ apps/block_scout_web/lib/block_scout_web/templates/address_signed/index.html.eex
@@ -7,7 +7,11 @@
     exchange_rate: @exchange_rate,
     coin_balance_status: @coin_balance_status,
     counters_path: @counters_path,
-    tags: @tags
+    tags: @tags,
+    validator_or_group_sum: @validator_or_group_sum,
+    voting_sum: @voting_sum,
+    locked_gold: @locked_gold,
+    vote_activated_gold: @vote_activated_gold
   %>
 
   <section data-page="blocks-signed">
