+++ apps/block_scout_web/lib/block_scout_web/templates/address_transaction/index.html.eex
@@ -2,7 +2,19 @@
 
   <% is_proxy = BlockScoutWeb.AddressView.smart_contract_is_proxy?(@address) %>
 
-  <%= render BlockScoutWeb.AddressView, "overview.html", address: @address, is_proxy: is_proxy, conn: @conn, exchange_rate: @exchange_rate, coin_balance_status: @coin_balance_status, counters_path: @counters_path, tags: @tags %>
+  <%= render BlockScoutWeb.AddressView, "overview.html",
+    address: @address,
+    is_proxy: is_proxy,
+    conn: @conn,
+    exchange_rate: @exchange_rate,
+    coin_balance_status: @coin_balance_status,
+    counters_path: @counters_path,
+    tags: @tags,
+    validator_or_group_sum: @validator_or_group_sum,
+    voting_sum: @voting_sum,
+    locked_gold: @locked_gold,
+    vote_activated_gold: @vote_activated_gold
+  %>
 
   <section data-page="address-transactions" id="txs">
     <div class="card">
